Organize AST modules.
